% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COPY_sparseSVM.R
\name{COPY_sparseSVM}
\alias{COPY_sparseSVM}
\title{Title}
\usage{
COPY_sparseSVM(X, y.train, ind.train, covar.train = NULL, alpha = 1,
  gamma = 0.1, nlambda = 100, lambda.min = if (nrow(X) > ncol(X)) 0.01
  else 0.05, lambda, screen = c("ASR", "SR", "none"), max.iter = 1000,
  eps = 1e-05, dfmax = ncol(X) + 1, penalty.factor = rep(1, ncol(X)),
  message = FALSE)
}
\arguments{
\item{X}{}

\item{y.train}{}

\item{ind.train}{}

\item{covar.train}{}

\item{alpha}{}

\item{gamma}{}

\item{nlambda}{}

\item{lambda.min}{}

\item{lambda}{}

\item{screen}{}

\item{max.iter}{}

\item{eps}{}

\item{dfmax}{}

\item{penalty.factor}{}

\item{message}{}
}
\description{
Title
}
